+++ tests/Unit/Generators/Factories/MethodFactoryTest.php
@@ -99,11 +99,13 @@ class MethodFactoryTest extends TestCase
             ->andReturn($doc);
 
         $this->statementFactory->shouldReceive('makeTodo')
-            ->with('Instantiate tested object to use it.')
-            ->andReturn(new TestStatement('/** @todo Instantiate tested object to use it. */'));
-        $this->statementFactory->shouldReceive('makeAffect')
-            ->with('foo', 'null')
-            ->andReturn(new TestStatement('$this->foo = null'));
+            ->with('Correctly instantiate tested object to use it.')
+            ->andReturn(new TestStatement('/** @todo Correctly instantiate tested object to use it. */'));
+        $this->statementFactory->shouldReceive('makeInstantiation')
+            ->with($class, Mockery::on(function ($value) {
+                return $value instanceof Collection && $value->isEmpty();
+            }))
+            ->andReturn(new TestStatement('$this->foo = new Foo()'));
 
         $method = $this->methodFactory->makeSetUp($class);
 
@@ -113,8 +115,8 @@ class MethodFactoryTest extends TestCase
         $this->assertSame([
             ['parent::setUp()'],
             [''],
-            ['/** @todo Instantiate tested object to use it. */'],
-            ['$this->foo = null'],
+            ['/** @todo Correctly instantiate tested object to use it. */'],
+            ['$this->foo = new Foo()'],
         ], $method->getStatements()->map(function (TestStatement $statement) {
             return $statement->getLines()->toArray();
         })->toArray());