+++ tests/Unit/Generators/Tests/Laravel/Channel/ChannelMethodFactoryTest.php
@@ -22,6 +22,7 @@ use Roave\BetterReflection\Reflection\ReflectionClass;
 use Roave\BetterReflection\Reflection\ReflectionMethod;
 use Roave\BetterReflection\Reflection\ReflectionProperty;
 use Tests\PhpUnitGen\Core\TestCase;
+use Tightenco\Collect\Support\Collection;
 
 /**
  * Class ChannelMethodFactoryTest.
@@ -105,11 +106,13 @@ class ChannelMethodFactoryTest extends TestCase
             ->andReturn(new TestImport('App\\User'));
 
         $this->statementFactory->shouldReceive('makeTodo')
-            ->with('Instantiate tested object to use it.')
-            ->andReturn(new TestStatement('/** @todo Instantiate tested object to use it. */'));
-        $this->statementFactory->shouldReceive('makeAffect')
-            ->with('foo', 'null')
-            ->andReturn(new TestStatement('$this->foo = null'));
+            ->with('Correctly instantiate tested object to use it.')
+            ->andReturn(new TestStatement('/** @todo Correctly instantiate tested object to use it. */'));
+        $this->statementFactory->shouldReceive('makeInstantiation')
+            ->with($class, Mockery::on(function ($value) {
+                return $value instanceof Collection && $value->isEmpty();
+            }))
+            ->andReturn(new TestStatement('$this->foo = new Foo()'));
         $this->statementFactory->shouldReceive('makeAffect')
             ->with('user', 'new User()')
             ->andReturn(new TestStatement('$this->user = new User()'));
@@ -119,8 +122,8 @@ class ChannelMethodFactoryTest extends TestCase
         $this->assertSame([
             ['parent::setUp()'],
             [''],
-            ['/** @todo Instantiate tested object to use it. */'],
-            ['$this->foo = null'],
+            ['/** @todo Correctly instantiate tested object to use it. */'],
+            ['$this->foo = new Foo()'],
             ['$this->user = new User()'],
         ], $method->getStatements()->map(function (TestStatement $statement) {
             return $statement->getLines()->toArray();